+++ pueue/src/bin/pueue.rs
@@ -2,7 +2,7 @@ use std::path::PathBuf;
 
 use anyhow::{bail, Context, Result};
 use clap::{CommandFactory, Parser};
-use clap_complete::{generate_to, shells};
+use clap_complete::{generate, generate_to, shells};
 use log::warn;
 use simplelog::{Config, ConfigBuilder, LevelFilter, SimpleLogger};
 
@@ -96,17 +96,50 @@ async fn main() -> Result<()> {
 /// [clap] is capable of creating auto-generated shell completion files.
 /// This function creates such a file for one of the supported shells and puts it into the
 /// specified output directory.
-fn create_shell_completion_file(shell: &Shell, output_directory: &PathBuf) -> Result<()> {
+fn create_shell_completion_file(shell: &Shell, output_directory: &Option<PathBuf>) -> Result<()> {
     let mut app = CliArguments::command();
     app.set_bin_name("pueue");
-    let completion_result = match shell {
-        Shell::Bash => generate_to(shells::Bash, &mut app, "pueue", output_directory),
-        Shell::Elvish => generate_to(shells::Elvish, &mut app, "pueue", output_directory),
-        Shell::Fish => generate_to(shells::Fish, &mut app, "pueue", output_directory),
-        Shell::PowerShell => generate_to(shells::PowerShell, &mut app, "pueue", output_directory),
-        Shell::Zsh => generate_to(shells::Zsh, &mut app, "pueue", output_directory),
+
+    // Output a completion file to a directory, if one is provided
+    if let Some(output_directory) = output_directory {
+        let completion_result = match shell {
+            Shell::Bash => generate_to(shells::Bash, &mut app, "pueue", output_directory),
+            Shell::Elvish => generate_to(shells::Elvish, &mut app, "pueue", output_directory),
+            Shell::Fish => generate_to(shells::Fish, &mut app, "pueue", output_directory),
+            Shell::PowerShell => {
+                generate_to(shells::PowerShell, &mut app, "pueue", output_directory)
+            }
+            Shell::Zsh => generate_to(shells::Zsh, &mut app, "pueue", output_directory),
+        };
+        completion_result.context(format!("Failed to generate completions for {shell:?}"))?;
+
+        return Ok(());
+    }
+
+    if let Some(output_directory) = output_directory {
+        let completion_result = match shell {
+            Shell::Bash => generate_to(shells::Bash, &mut app, "pueue", output_directory),
+            Shell::Elvish => generate_to(shells::Elvish, &mut app, "pueue", output_directory),
+            Shell::Fish => generate_to(shells::Fish, &mut app, "pueue", output_directory),
+            Shell::PowerShell => {
+                generate_to(shells::PowerShell, &mut app, "pueue", output_directory)
+            }
+            Shell::Zsh => generate_to(shells::Zsh, &mut app, "pueue", output_directory),
+        };
+        completion_result.context(format!("Failed to generate completions for {shell:?}"))?;
+
+        return Ok(());
+    }
+
+    // Print the completion file to stdout
+    let mut stdout = std::io::stdout();
+    match shell {
+        Shell::Bash => generate(shells::Bash, &mut app, "pueue", &mut stdout),
+        Shell::Elvish => generate(shells::Elvish, &mut app, "pueue", &mut stdout),
+        Shell::Fish => generate(shells::Fish, &mut app, "pueue", &mut stdout),
+        Shell::PowerShell => generate(shells::PowerShell, &mut app, "pueue", &mut stdout),
+        Shell::Zsh => generate(shells::Zsh, &mut app, "pueue", &mut stdout),
     };
-    completion_result.context(format!("Failed to generate completions for {shell:?}"))?;
 
     Ok(())
 }